#46681 get OS information from a Bean instead of a global function
