+++ cdr-client-service/src/main/kotlin/com/swisscom/health/des/cdr/client/handler/CdrApiClient.kt
@@ -5,7 +5,6 @@ import com.swisscom.health.des.cdr.client.common.Constants.EMPTY_STRING
 import com.swisscom.health.des.cdr.client.common.getRootestCause
 import com.swisscom.health.des.cdr.client.config.CdrClientConfig
 import com.swisscom.health.des.cdr.client.config.HttpServerErrorException
-import com.swisscom.health.des.cdr.client.getOsInfo
 import io.github.oshai.kotlinlogging.KotlinLogging
 import okhttp3.Headers
 import okhttp3.MediaType.Companion.toMediaType
@@ -38,7 +37,8 @@ internal class CdrApiClient(
     private val httpClient: OkHttpClient,
     @param:Qualifier("retryIoAndServerErrors")
     private val retryIOExceptionsAndServerErrors: RetryTemplate,
-    private val objectMapper: ObjectMapper
+    private val objectMapper: ObjectMapper,
+    private val osInformation: String
 ) {
 
     @Suppress("LongMethod", "CyclomaticComplexMethod")
@@ -363,7 +363,7 @@ internal class CdrApiClient(
             this[AZURE_TRACE_ID_HEADER] = traceId
             this[CLIENT_TYPE_HEADER] = clientType ?: "cdr-client"
             this[CLIENT_VERSION_HEADER] = clientVersion ?: "unknown"
-            this[CLIENT_OS_HEADER] = getOsInfo()
+            this[CLIENT_OS_HEADER] = osInformation
             this.build()
         }
     }
